ReConstruct some abstract command.
