+++ ratis-shell/src/main/java/org/apache/ratis/shell/cli/sh/command/AbstractParentCommand.java
@@ -24,19 +24,28 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
-public abstract class AbstractParentCommand extends AbstractRatisCommand{
+public abstract class AbstractParentCommand implements Command {
   private final Map<String, Command> subs;
 
   public AbstractParentCommand(Context context, List<Function<Context, AbstractRatisCommand>> subCommandConstructors) {
-    super(context);
     this.subs = Collections.unmodifiableMap(subCommandConstructors.stream()
         .map(constructor -> constructor.apply(context))
         .collect(Collectors.toMap(Command::getCommandName, Function.identity(),
         (a, b) -> {
           throw new IllegalStateException("Found duplicated commands: " + a + " and " + b);
           }, LinkedHashMap::new)));
+  }
+
+  public AbstractParentCommand(List<Supplier<AbstractCommand>> subCommands) {
+    this.subs = Collections.unmodifiableMap(subCommands.stream()
+        .map(subCommand -> subCommand.get())
+        .collect(Collectors.toMap(Command::getCommandName, Function.identity(),
+        (a, b) -> {
+            throw new IllegalStateException("Found duplicated commands: " + a + " and " + b);
+            }, LinkedHashMap::new)));
   }
 
   @Override
